Updated Delete Inventory Logic to check if an inventory copy is available. Updated model relation between History and Inventory to allow cascading removal of history upon Inventory deletion. Updated controller logic to present more information to the user.

+++ src/main/java/org/bitmaxsystems/oop2library/services/inventoryService/DeleteInventoryCopyService.java
@@ -1,6 +1,10 @@
 package org.bitmaxsystems.oop2library.services.inventoryService;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.bitmaxsystems.oop2library.controllers.AdministrativeInventoryController;
 import org.bitmaxsystems.oop2library.models.inventory.Inventory;
+import org.bitmaxsystems.oop2library.models.inventory.states.InventoryStateEnum;
 import org.bitmaxsystems.oop2library.repository.GenericRepository;
 
 public class DeleteInventoryCopyService {
@@ -8,11 +12,28 @@ public class DeleteInventoryCopyService {
     private final GenericRepository<Inventory> inventoryRepository =
             new GenericRepository<>(Inventory.class);
 
-    public void delete(Inventory inventory) {
-        if (inventory == null) {
-            throw new IllegalArgumentException("Inventory copy cannot be null.");
-        }
+    private static final Logger logger =
+            LogManager.getLogger(DeleteInventoryCopyService.class);
 
-        inventoryRepository.delete(inventory);
+    public void delete(Inventory inventory) {
+
+       try {
+
+
+           if (inventory == null) {
+               throw new IllegalArgumentException("Inventory copy cannot be null.");
+           }
+
+           if (inventory.getState().getStateEnum() != InventoryStateEnum.AVAILABLE) {
+               throw new IllegalArgumentException("Inventory must be available before deletion");
+
+           }
+
+           inventoryRepository.delete(inventory);
+           logger.info("Inventory copy successfully deleted");
+       } catch (Exception e) {
+           logger.error(e);
+           throw e;
+       }
     }
 }
